+++ scripts/test-grammar.sh
@@ -25,6 +25,7 @@ esac
 source "$(dirname "$0")/grammar-test-lib.sh"
 
 grammar_generate
+grammar_build_wasm
 grammar_lint
 grammar_corpus
 grammar_highlight